backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "engsoftcris", email = "cristiano.tobias40@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.1,<7.0.0",
    "djangorestframework>=3.16.1,<4.0.0",
    "psycopg2-binary>=2.9.11,<3.0.0",
    "pillow>=12.1.0,<13.0.0",
    "djangorestframework-simplejwt>=5.5.1,<6.0.0",
    "django-cors-headers>=4.9.0,<5.0.0",
    "social-auth-app-django>=5.7.0,<6.0.0",
    "whitenoise (>=6.11.0,<7.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest>=9.0.2,<10.0.0",
    "pytest-django>=4.11.1,<5.0.0",
    "factory-boy>=3.3.3,<4.0.0",
    "faker>=40.1.2,<41.0.0",
    "black>=26.1.0,<27.0.0"
]
